Simulate events with non-null myo::Myo*
implemented by getting N myos, where N is the number of myos in the
event queue, and mapping each MyoEvent::myo_index to a myo obtained
with Hub::waitForMyo

fixes #10

+++ src/EventRecorder.cpp
@@ -130,7 +130,7 @@ void EventRecorder::onWarmupCompleted(myo::Myo* myo, uint64_t timestamp,
                                                                  warmupResult));
 }
 
-int EventRecorder::getOrAddMyoIndex(myo::Myo* myo) {
+size_t EventRecorder::getOrAddMyoIndex(myo::Myo* myo) {
   if (myo_indicies_.count(myo) != 0) return myo_indicies_[myo];
   return myo_indicies_[myo] = myo_indicies_.size();
 }
